refactor: 큐레이션 등록 모달의 상태 및 로직 훅으로 분리
